scrapingv2/backend/requirements.txt
# Core LLM & Data Structuring
google-generativeai==0.8.3
pydantic==2.8.2
pydantic-settings==2.4.0

# Academic Scraping
google-search-results>=2.4.2

# Scientific Evaluation Metrics
scikit-learn>=1.3.0

# Environment Management
python-dotenv==1.0.1